challenges/cryptography/puzzle solving/challenge/app/requirements.txt
Flask>=2.2.2
pycryptodome==3.21.0
